+++ tests/test_default.rs
@@ -67,7 +67,6 @@ mod suffix_reverse;
 mod unicode;
 mod word_boundary;
 mod word_boundary_unicode;
-mod error_messages;
 
 #[test]
 fn disallow_non_utf8() {